feat/product-service
 "Fix error codes in category and product service; update product service handleQueueOperation"

The diff includes changes to the category and product service files. In the category service, the error codes for certain if statements have been updated to return a 400 Bad Request instead of a 404 Not Found. In the product service, the error codes for certain if statements have been updated to return a 401 Unauthorized instead of a 403 Forbidden. Additionally, the handleQueueOperation function in the product service has been updated to include the retrieval of product information and return it in the success response.